backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg[binary]
python-dotenv
PyJWT
bcrypt

